Modified Mega2560 bootloaders to deal with intel hex records that span two flash pages.
Made changes and retry logic to the Arduino sketch to improve performance
while waiting for the target processor to complete flash writes. This was
done to prevent incorrect detection of flash write failures that were caused
by the dynamic nature of flash programming times from page to page.

The reset interrupt vector record is now written last. This has the effect
of enabling the loaded applicaion at the end of the cycle rather than at
the begginning.